requirements.txt
Django==2.0.4
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
pkg-resources==0.0.0
pytz==2018.4

